# crates/storage-api/openapi-storage-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Storage API",
    "description": "",
    "license": {
      "name": "BUSL-1.1",
      "identifier": "BUSL-1.1"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/storage": {
      "post": {
        "tags": [
          "crate::routes::dynamodb"
        ],
        "summary": "Execute a DynamoDB-compatible storage operation.",
        "description": "Sends a DynamoDB-compatible JSON-RPC request through the AuxFn storage gateway using the `x-amz-target` header to select the operation. The aggregated internal-private `/storage` surface accepts either an authenticated bearer session or an `x-api-key` service token bound to the `storage-gateway` service, while non-DynamoDB helper traffic stays on separate internal routes.",
        "operationId": "dynamodb_endpoint",
        "parameters": [
          {
            "name": "x-amz-target",
            "in": "header",
            "description": "DynamoDB operation target (e.g., DynamoDB_20120810.CreateTable)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "description": "Optional service token bound to the storage-gateway service when the caller is a machine principal",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "DynamoDB operation successful",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Wrong machine credential type or service binding",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Admission overloaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-aux-audience": "internal_private",
        "x-aux-internal": true,
        "x-aux-authn-classification": "session_or_service_token"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "__type",
          "message"
        ],
        "properties": {
          "__type": {
            "type": "string",
            "description": "AWS-compatible JSON error field name. DynamoDB and SQS JSON protocol\nerrors use `__type`. Machine-readable API error type.",
            "example": "ValidationException"
          },
          "documentation_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional documentation link for operator or developer remediation.",
            "example": "/docs/explanations/error-code-catalog"
          },
          "message": {
            "type": "string",
            "description": "AWS-compatible JSON error field name. DynamoDB and SQS JSON protocol\nerrors use `message`. Human-readable error message.",
            "example": "One or more parameter values were invalid."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Request correlation identifier echoed in the response body.",
            "example": "018f1f61-2a6f-7ac3-b9b6-7f65bb2d91fd"
          },
          "retry_after_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Optional retry budget for rate-limited responses.",
            "example": 2,
            "minimum": 0
          }
        }
      },
...
    }
  }
}